Add new ListingsCoverImage model; update schema
In order to simplify the process for adding images to Listings in the
Django backend, I created a new intermediary model and table ListingsCoverImage.
Its one-to-one relationships with the Listings model also represents
creates an abstraction that, I think, shows the relationship between Listings
and Image better than holding a direct reference.

Also updated the schema in order to reflect this change..