Revise CSRF validation to occur when executing mutations
Created a new decorator to be used whenever we perform a mutation.

The reason for this is because web crawlers were unable to receive
and validate a CSRF token when generating pages while the validation
was being perfomed on the entire GraphQL view. Also, it just makes
sense to only perform the validation when actually sending data to the API,
not for simple scenarios like GETs.