Work in Progress - Implement refactor of profile image upload via new Storage class
In order to allow for more fine-grained saving of responsive images,
I wrote a new storage class that will allow for automatic resizing of
images to 1x, 2x, 3x, and 4x.

Everything seems to also be working with the native Django admin,
which will also likely prevent lots of bugs in the future, as doing
all of this processing via the GraphQL schema was messy and
prone to error.

The new approach still needs to be tested thoroughly and synced
with the front end. Also, it looks like processing the images could
be a lengthy process depending on the size of the image. Also, saving
13 versions of every image seems excessive, so will see if I can mitigate
this somehow.

In any case, the functionality now exists and should be able to adapt
the approach to fit whatever strategy I end up taking.

Next steps are to do thorough testing, determine the correct strategy
for handling retina images, clean up code and sync with the front end.