Clean up code for creating image thumbnails
Modified the catalog storage class so that the image saving
for the thumbnails is more robust. Also implemented the deletion
of existing profile images when updating images via GraphQL.

Importantly, when generating thumbnails, the image file name must include
the '-1x' string so that subsequent sizes can be swapped into the name.