Remove 'slug' from Listing model; create slug and image migrations
Storing a slug on the Listing model is no longer necessary, as
I want to reduce cognitive overhead for users during listing management
as well as create a future-proof URL scheme for the front end.

The slug is now simply derived from the listing title, and the frontend
uses the lising id exlusively to retrieve data.

This change has been reflected in the current commit.