Modify listing management to use id instead of slug; continue to develop UpdateListing mutation
We are now using id instead of slug for the frontend to reduce
cognitive overhead for users and make the entire system more future-proof.
Slug is now derived from the title, but it's only used on the frontend
as a 'nice to have' for SEO purposes.

Other than that, added a few fields and modifications to related to
continuing development of the update listing functionality.