Overhaul thumbnail creation; revise ListingCoverImage; add DeleteListing mutation
Make big changes to thumbnail creation in order to make the
the generation of the thumbnails and their responsive sizes
more stable, as well as improve server response time when uploading
images.

Only the saving the 'original' image attribute is done during the main
thread now - the creation of thumbnails has been offloaded to a separate
thread.

In addition, all the saving / deletion cleanup logic related to images
has been removed from the schema and instead placed on the ListingImage
model, which, along with tasks.py, constants.py, and signals.py, now
makes image handling more scalable.

Importantly, all of the logic works via both the GraphQL endpoint and the
Django admin. I still need to do some cleanup around this - still need
to modify the admin slightly to prevent anyone from mistakenly causing
data integrity issues.

Worth noting is that all responsive image sizes are now derived from the
image saved in the 'original' attribute.

Early on, I made an unwise decision to create a custom id for ListingCoverImage,
which was causing headaches with the new image handling. So I manually
created an autofield that functions exactly like the Django default using
a custom migration.

Note that I needed to create some new configuration settings to allow
us to process signals, which now perform cleanup on the deletion of a listing.

Next step is to add validation to the flow for byline creation, do thorough
testing of creation and updating of users and listings, do any code cleanup
and move towards deployment.