Add requester and is_confirmed to byline models; add ConfirmByline mutation
Needed to make some modifications to ListingCreatorByline and
ListingCollaboratorByline in order to accommodate the new 'requester'
field. Namely, the many-to-many relationship with User requires the
addition of the 'through fields' attribute because there are now
two instances of User on the byline models.

The other changes are straightforward - a new mutation that
checks for authorization before updating or deleting a byline.

May need to modify this mutation - not sure whether it's better
to use the listing id or byline id in order to access and update.

In any case, next steps are to move onto a preliminary deployment
for testing.

Note that this feature isn't entirely ready - would also like to
generate an email whenever a new byline is requested. After that though,
should come code cleanup in preparation for the alpha deployment.