Fix recreation / duplication of byline bug
For most fields on listings, we can dump the existing attributes on the
listing for every update and recreate the fields using whatever
data we receive from the frontend.

However, when it comes to bylines, we can't do that, because we need to
preserve the is_confirmed attribute between form submissions.

To address this, added some extra handling / looping logic to deal with this.
It's not the cleanest implementation - will think about fixing it
in the future.