Remove deprecated Image model; simplify model imports; consolidate image functionality
Moved UserProfileImage to use the image functionality and thumbnail creation used
for the listings. Also, introduced a new ArticleFeaturedImage model that makes
use of the functionality as well.

I abstracted some helper methods and created abstract base classes
to avoid repetition and clean things up overall.

Since we're no longer using the Image model, I removed it. A sequence of migrations
should make it so that this doesn't cause problems when I move to deploy everything
(hopefully).

NOTE: A big thing that still needs to be done is make a method that will clean up unused
profile and article images as well (right now, the clean_bucket method inside of
catalog.tasks only works for listings).

Next steps are to continue consolidating and cleaning up the codebase as I move
toward an alpha release.