Add ability to delete collaborator bylines in UpdateUser mutation
Collaborator bylines can now be deleted from the edit user profile page.
Note: This functionality is specific to *collaborator* bylines. Creator
bylines must be removed from the edit listing page, as normal.

The reason for this is otherwise it's possible to end up with inconsistencies,
such as listings having no creators, and while possible to implement this,
I feel like it's low priority and will require some thought in order
to get right.

Next steps are to continue polishing code and user experience.