Modify schema to accomodate new Next image component on front end
We may be able to remove the responsive image generation on the
backend. It turns out the Next framework has devised robust handling
for this.

Will push this up to test. If it works, next steps are to continue
implementing on the frontend and potentially eliminate extraneous
backend code that results from the change.