Modify confirm byline mutation to return user object and update store
Apollo should be able to update its cache automatically based on information
we return after a mutation.

With this in mind, I modified Confirm Byline in order to refresh the user's
page upon completion.

Unfortunately this doesn't seem to be working as expected - will likely need
to devote more time to get it working correctly. For now, forcing a hard
refresh and refetch on the front end will do.

Next steps are to continue troubleshooting production / caching issues.