Add callbacks for deleting articles and listings; fix admin errors
I realized that ContentThreads, Threads, and Comments weren't getting
deleted as they should when articles and listings are deleted. So that
has now been fixed. Also, the admin was firing errors when certain
elements were no longer in existence, so changed getattr() methods in
a few places to hasattr() instead.