Fix image upload bugs
Realized that not all images were being saved with correct ContentType,
so fixed that inside of ThumbnailImageStorage.

Realized also that article images were not being thumbnailed correctly,
so hardcoded thumbnail values into the ThumbnailImageStorage for now,
though can likely implement this better in the near future.

Also, made sure that article image names now get sanitized before
executing uploads so that we're not dealing with spaces and special
characters on the frontend.